Move implementations of fill_in_forest and get_cyclic_snarl_intervals into the cpp file
